fix(lpa-routes): handle lpa-dashboard auth + feature check in one location
